Restructured project again.
