--- src/components/about/components/bullet-point/index.tsx
@@ -1,1 +1,0 @@
-export { default as BulletPoint } from './bullet-point';
