--- src/components/about/components/index.tsx
@@ -1,1 +1,0 @@
-export { BulletPoint } from './bullet-point'
