+++ src/components/app/style.ts
@@ -1,0 +1,16 @@
+/**
+ * The styles for theApp component.
+ */
+
+ import { makeStyles } from "@material-ui/core/styles";
+
+
+const useStyles = makeStyles({
+  root: {
+    height: "100vh",
+    display: "flex",
+  },
+});
+
+
+export { useStyles };
