+++ src/components/index.tsx
@@ -1,4 +1,4 @@
-export { About } from './about';
-export { LoginForm } from './login-form';
-export { RegisterForm } from './register-form';
-export { RecoverPasswordForm } from './recover-password-form';
+export { default as About } from './about';
+export { default as LoginForm } from './login-form';
+export { default as RegisterForm } from './register-form';
+export { default as RecoverPasswordForm } from './recover-password-form';
