+++ src/components/lock-text/style.ts
@@ -1,0 +1,16 @@
+/**
+ * The styles for the LockText component.
+ */
+
+import { makeStyles } from "@material-ui/core/styles";
+
+
+const useStyles = makeStyles(theme => ({
+  avatar: {
+    margin: theme.spacing(1),
+    backgroundColor: "white",
+  },
+}));
+
+
+export { useStyles };
