+++ src/index.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { ThemeProvider, createMuiTheme } from '@material-ui/core/styles';
-import { App } from './components/app';
 import { themeOptions } from './theme-options';
+import App from './components/app';
+
 
 const theme = createMuiTheme(themeOptions);
+
 
 ReactDOM.render(
   <React.StrictMode>
