--- src/views/home/Home.tsx
@@ -1,5 +1,0 @@
-/**
- * The website homepage.
- */
-
-export {}
