+++ src/views/register/style.ts
@@ -1,0 +1,15 @@
+/**
+ * The styles for the Register page.
+ */
+
+import { makeStyles } from "@material-ui/core/styles";
+
+
+const useStyles = makeStyles(theme => ({
+  gridItem: {
+    padding: theme.spacing(2),
+  },
+}));
+
+
+export { useStyles };
